ai/requirements.txt
# train model
tensorflow>=2.4.1
scikit-learn
xgboost
imblearn
pandas

# server
fastapi
fastapi-socketio
uvicorn[standard]
pydantic
xgboost
scikit-learn